Category endpoint cream (#19)
* Finish get category income & expense

* Add get category income and expense endpoint

* ADD endpoint transaction post

+++ backend/cmd/models/dto/profile/profile.go
@@ -1,8 +1,8 @@
 package profile
 
 type Profile struct {
-	Email              string `json:"email"`
-	UserName           string `json:"user_name"`
-	PathProfilePicture string `json:"path_profile_picture"`
-	Balance            int64  `json:"balance"`
+	Email              string  `json:"email"`
+	UserName           string  `json:"user_name"`
+	PathProfilePicture string  `json:"path_profile_picture"`
+	Balance            float64 `json:"balance"`
 }
